Use 2 worker threads per POV-Ray invocation. ...
By default, POV-Ray will assume we want a number of threads equivalent
to the number of CPUs. That results in overall slowness and excessive
thread-spawning when used in conjunction with a job dispatcher like GNU
parallel.

This may not be a good choice when processing large files that may
benefit from several cores, but in my testing, images from the Theta V
are processed more quickly with fewer threads.