Redo our overhaul of exiftool invocation. ...
We want to use -m to squelch the warning about MakerNotes offset.
-F doesn't actually fix it, and the notes don't copy correctly.

We want to specify that we're copying -all:all. This ensures that we
include some Ricoh-specific tags that were getting left behind.

We want to set the pitch and roll to zero, not remove them altogether.
Arguably, the Ricoh Pitch and Ricoh Roll values should also be zeroed,
but leaving them for now.

The ProcessingSoftware tag seems like a  more semantically meaningful
place for our mark than the Software tag.